Update INDI-event.md
Military Service is currently listed as an Attribute, where as this should be an event.  GEDCOM-X and other applications list multiple new military service based tags for inclusion into new versions of GEDCOM.  this proposal suggests that adding a single event with support for enumerated `MILT.TYPE` values that provide for common military events that can be translated to all languages, with an additional final enumeration of 'other' and a `PHRASE` tag.